Conbined SchoolUtils to Schols controller since it is the only controller for that endpoint now.
